Refactoring: new file functions.py to store functions for film response and fitting.
