+++ tests/unit/fixtures/CAL/for_test_plot_central.py
@@ -1,4 +1,4 @@
-from Dosepy.calibration import CalibrationLUT, _get_dose_from_fit
+from Dosepy.calibration import LUT, _get_dose_from_fit
 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -6,7 +6,7 @@ from Dosepy.image import load
 
 path_file = "film20240620_002.tif"
 img = load(path_file)
-cal = CalibrationLUT(img)
+cal = LUT(img)
 cal.set_central_rois((8, 8), show = True)
 cal.set_doses([0, 1, 2, 4, 6.5, 9.5])
 #cal.compute_central_lut(filter = 3)
