+++ tests/unit/fixtures/CAL/for_test_yaml.py
@@ -1,7 +1,12 @@
-from Dosepy.calibration import CalibrationLUT
+from Dosepy.calibration import LUT
 import matplotlib.pyplot as plt
 
-cal = CalibrationLUT.from_yaml_file("calibration.yaml")
+cal = LUT.from_yaml_file("calibration.yaml")
 
-cal.plot_fit(fit_type="polynomial", position=0, channel="red")
+cal.plot_fit(
+    fit_type="polynomial",
+    position=0,
+    channel="red"
+    )
+
 plt.show()
